Remove likely multi-nucleotide mutation events Function that removes DNP that are actually labeled as single nucleotide variants.

Remove likely multi-nucleotide mutation events

Function that removes DNP that are actually labeled as single nucleotide variants.

DNP_TNP_remover(MAF, delete_recur = F)

Arguments

MAF
MAF file, with relevant fields consisten with the GDC MAF specification (https://docs.gdc.cancer.gov/Data/File_Formats/MAF_Format/).

delete_recur
T/F statement about whether to remove recurrent (02) tumors.