Gene-level epistasis

Calculate selection intensity under an assumption of pairwise gene-level epistasis. Selection at the gene level is assumed to act through all specified variants (see options), with the selection for each gene's variants allowed to vary based on the mutational status of the other gene's variants.

Usage

ces_gene_epistasis(
  cesa = NULL,
  genes = NULL,
  variants = NULL,
  samples = character(),
  run_name = "auto",
  cores = 1,
  conf = 0.95,
  return_fit = FALSE
)

Arguments

cesa

CESAnalysis object

genes

Vector of gene names; SIs will be calculated for all gene pairs. Alternatively, a list of gene pairs (2-length character vectors) to run just the given pairings.

variants

Which variants to include in inference for each gene. Either "recurrent" for all variants present in two or more samples (across all MAF data), "nonsilent" for nonsynonymous coding variants and variants in essential splice sites, or a data.table containing all variants to include (as returned by select_variants() or by subsetting[CESAnalysis\]$variants). For noncoding variants with multiple gene annotations, the one listed in the "gene" column is used. In the recurrent method, nearby noncoding variants may be included.

samples

Which samples to include in inference. Defaults to all samples. Can be a vector of Unique_Patient_Identifiers, or a data.table containing rows from the CESAnalysis sample table.

run_name

Optionally, a name to identify the current run.

cores

number of cores for parallel processing of gene pairs

conf

Confidence interval size from 0 to 1 (.95 -> 95%). NULL skips calculation, which may be helpful to reduce runtime when analyzing many gene pairs.

return_fit

Temporary option for whether to embed fitted models in a "fit" attribute associated with the epistasis results table. This argument will be replaced soon with friendlier support for accessing fitted models.

Value

CESAnalysis with epistasis analysis results added

Details

Only samples that have coverage at all included sites in both genes can be included in the inference since samples lacking full coverage may or may not have mutations at the uncovered sites.