Calculate selection intensity for single-nucleotide variants and amino acid changes

Calculate selection intensity for single-nucleotide variants and amino acid changes

ces_snv(
  cesa = NULL,
  genes = "all",
  cores = 1,
  include_genes_without_recurrent_mutations = F,
  find_CI = T
)

Arguments

cesa
CESAnalysis object

cores
number of cores to use

include_genes_without_recurrent_mutations
default false; will increase runtime and won't find anything interesting

gene
which genes to calculate effect sizes within; defaults to all genes with recurrent mutations in data set

Value

CESAnalysis object with selection results added for the chosen analysis