ml_objective for epistasis calculation

Objective function that we will be optimizing in order to find the site specific selection intensity that maximizes the likelihood of each tumor having a mutation or not, where the mutation rates are site and tumor specific.

ml_objective_epistasis(
  gamma,
  MAF_input1,
  MAF_input2,
  all_tumors,
  variant1,
  variant2,
  specific_mut_rates1,
  specific_mut_rates2
)

Arguments

gamma
A selection intensity at which to calculate the likelihood

MAF_input1
A data frame that includes columns "Unique_patient_identifier", "Gene_name", and "unique_variant_ID_AA"

MAF_input2
A data frame that includes columns "Unique_patient_identifier", "Gene_name", and "unique_variant_ID_AA"

all_tumors
A list of all the tumors we are calculating the likelihood across

variant1
one variant in epistatic analysis

variant2
the other

specific_mut_rates1
site and tumor specific mutation rates (matrix: rows = tumors, columns = variants; produced by mutation_rate_calc)

specific_mut_rates2
site and tumor specific mutation rates (matrix: rows = tumors, columns = variants; produced by mutation_rate_calc)

Value

A log likelihood value