cancereffectsizeR wrapper for deconstructSigs

This function gets called internally by trinuc_mutation_rates() for each tumor in a CESAnalysis, accepting its a data.frame of mutation counts and returning deconstructSigs output.

run_deconstructSigs(
  tumor_trinuc_counts,
  signatures_df,
  signatures_to_remove,
  artifact_signatures,
  tri.counts.method
)

Arguments

tumor_trinuc_counts
one-row data.frame of trinuc variant counts (in deconstructSigs order) for one tumor

signatures_df
data.frame of signatures (see COSMIC v3 signatures included with package for format)

signatures_to_remove
names of signatures in signatures_df to keep out of deconstructSigs and assign zero weights

artifact_signatures
names of signatures that should be treated as experimental (e.g., sequencing) artifacts (these have weights calculated but are then normalized out when calculating true trinuc proportions)

tri.counts.method
exome/genome trinucleotide content normalization argument to pass to deconstructSigs (see its docs)

Value

a list with 2 items: raw_sig_output (exactly what comes out of deconstructSigs) and adjusted_sig_output, which contains signature weights (possibly adjusted with artifact accounting) and expected relative trinuc mutation rates based on signature weights

Details

Since this funciton implements artifact accounting and signature removal, you may find it useful for your own mutational signature analyses. However, you may be able to just use trinuc_mutation_rates().