Select and filter variants

This function helps you find and view variant data from your CESAnalysis's MAF data and mutation annotation tables. By default, almost all amino-acid-changing mutations and noncoding SBS are returned. You can apply a series of filters to restrict output to certain genes or genomic regions, require a minimum variant frequency in MAF data, and/or specify exact variant_ids to return.

Usage

select_variants(
  cesa,
  genes = NULL,
  min_freq = 0,
  variant_ids = NULL,
  type = NULL,
  gr = NULL,
  variant_position_table = NULL,
  padding = 0
)

Arguments

cesa

CESAnalysis with MAF data loaded and annotated (e.g., with load_maf())

genes

Filter variants to specified genes.

min_freq

Filter out variants with MAF frequency below threshold (default 0). Note that variants that are not in the annotation tables will never be returned. Use add_variants() to include variants absent from MAF data in your CESAnalysis.

variant_ids

Filter out any variants besides those specified. (Variants specified here can still get removed by other filters.)

type

Vector of variant types to include, such as "sbs" or "aac".

gr

Filter out any variants not within input GRanges +/- padding bases.

variant_position_table

Filter out any variants that don't intersect the positions given in chr/start/end of this table (1-based closed coordinates). Typically, the table comes from a previous select_variants call and can be expanded with padding. (Gritty detail: Amino acid substitutions get special handling. Only the precise positions in start, end, and center_nt_pos are used. Otherwise, coding changes that span splice sites would capture all the intronic sites.)

padding

add +/- this many bases to every range specified in gr or variant_position_table (stopping at chromosome ends, naturally).

Value

A data.table with info on selected variants (see details)

Details

Only variants that are annotated in the CESAnalysis can be returned. To view more annotations, such as for variants absent from the MAF data, you must first call add_variants() to add them to the CESAnalysis.

Note that while intergenic SBS have their nearest genes annotated in the SBS tables, these variants will not be captured by the genes filter of this function.

Definitions of some less self-explanatory columns:

variant_name: In a coding variant, gene and protein change on the (MANE) canonical transcripts, such as "BRAF V600E". For coding changes reported on other transcripts, the protein ID is included: "POLH W415C (ENSP00000361300.1)". With older reference data sets (ces.refset.hg19, versions of ces.refset.hg38 < 1.3, and any custom reference data set that doesn't have complete information on canonical transcripts), the variant name is a shortening of the variant_id.

start/end: lowest/highest genomic positions overlapping variant

variant_id: unique IDs for variants given the associated genome assembly version and the transcript data

ref/alt: genomic reference and alternate alleles (for genomic variants; NA for AACs)

gene: the affected gene in AACs; for SBS, the overlapping gene (or an arbitrary gene if more than one overlaps), or the nearest gene for intergenic SBS

strand: for AACs, 1 if the reference sequence strand is the coding strand; -1 otherwise

essential_splice: Variant is 1,2 bp upstream or 1,2,5 bp downstream of an annotated splice position (edge case: if an SBS has multiple gene/transcript annotations, this doesn't say which one it's essential for)

intergenic: variant does not overlap any coding regions in the reference data

trinuc_mut: for SBS, the reference trinucleotide context, in deconstructSigs notation

coding_seq: coding strand nucleotides in order of transcription

center_nt_pos: regardless of strand, start/end give positions of two out of three AAC nucleotides; this gives the position of the center nucleotide (maybe useful if the AAC spans a splice site)

multi_anno_site: T/F whether variant has multiple gene/transcript/AAC annotations

maf_prevalence: number of occurrences of the variant in MAF data

samples_covering: number of MAF samples with sequencing coverage at the variant site