buildref

Function to build a RefCDS object from a reference genome and a table of transcripts. The RefCDS object has to be precomputed for any new species or assembly prior to running dndscv. This function generates an .rda file that needs to be input into dndscv using the refdb argument. Note that when multiple CDS share the same gene name (second column of cdsfile), the longest coding CDS will be chosen for the gene. CDS with ambiguous bases (N) will not be considered.

build_RefCDS(
  cds_data,
  genome,
  numcode = 1,
  excludechrs = NULL,
  onlychrs = NULL,
  useids = F,
  cores = 1
)

Arguments

cds_data
data from biomaRt for the genome assembly of interest

genome
a BSgenome object corresponding to the genome assembly of interest

numcode
NCBI genetic code number (default = 1; standard genetic code). To see the list of genetic codes supported use: ? seqinr::translate

excludechrs
Vector or string with chromosome names to be excluded from the RefCDS object (default: no chromosome will be excluded). The mitochondrial chromosome should be excluded as it has different genetic code and mutation rates, either using the excludechrs argument or not including mitochondrial transcripts in cdsfile.

onlychrs
Vector of valid chromosome names (default: all chromosomes will be included)

useids
Combine gene IDs and gene names (columns 1 and 2 of the input table) as long gene names (default = F)

cores
how many cores to use for parallel computations (requires parallel package)

Details

Based on the buildref function in Inigo Martincorena's package dNdScv.