Calculate selection intensity for single-nucleotide variants and amino acid changes

Calculate selection intensity for single-nucleotide variants and amino acid changes

ces_variant(
  cesa = NULL,
  variants = select_variants(cesa, min_freq = 2),
  model = "sswm",
  group_ordering = NULL,
  custom_lik_args = list(),
  hold_out_same_gene_samples = is(variants, "data.table"),
  cores = 1,
  conf = 0.95
)

Arguments

cesa
CESAnalysis object

variants
Variant table from select_variants(), or a CompoundVariantSet from define_compound_variants(). Defaults to all recurrent noncoding SNVs and (SNV-derived) coding mutations, where recurrent means appearing in at least two samples in the MAF data set.

model
"sswm" or "sswm_sequential" to use built-in models of selection, or supply a custom function factory (see details).

group_ordering
for models (like sswm_sequential) that assume different groups of samples are from sequential tumor progression states, a vector or list giving the ordering of CESAnalysis groups. Examples: c("Primary", "Metastatic") is a two-state ordering. Lists allow groups to share progression states: list("A", c("B", "C"), "D") means that group A is first, groups B and C belong to the same intermediate state, and group D is last.

custom_lik_args
Extra arguments, given as a list, to pass to custom likelihood functions.

hold_out_same_gene_samples
When finding likelihood of each variant, hold out samples that lack the variant but have other mutations in the same gene. By default, TRUE when running with single variants, FALSE with compound variants.

cores
number of cores to use

conf
selection intensity confidence interval width (NULL skips calculation, speeds runtime)

Value

CESAnalysis object with selection results added for the chosen analysis