Use dNdScv with tissue-specific covariates to calculate gene-level mutation rates

This function calculates gene-level neutral mutation rates based on counts of nonsynonymous and synonymous mutations per gene under the dNdScv package's model, as described in Martincorena et al. (https://doi.org/10.1016/j.cell.2017.09.042).

gene_mutation_rates(cesa, covariates = NULL, save_all_dndscv_output = FALSE)

Arguments

cesa
CESAnalysis object

covariates
Name of gene mutation rate covariates to use (run list_ces_covariates() to see choices). For hg19 only, you can also use "default" for dNdScv's non-tissue-specific covariates. If no covariates are available, set NULL to run without.

save_all_dndscv_output
default false; when true, saves all dndscv output, not just what's needed by CES (will make object size very large)

Value

CESAnalysis object with gene-level mutation rates calculated