Load MAF-formatted somatic mutation data

Load MAF data from a text file or data.frame/data.table into a CESAnalysis object. If column names don't match MAF format specifications (Chromosome, Start_Position, etc., with Tumor_Sample_Barcode used as the sample ID column), you can supply your own column names. When your CESAnalysis includes chronological tumor progression states (e.g., pre- vs. post-treatment, stages 1-4, primary vs. metastatic), specify "progression_col". By default, data is assumed to be derived from whole-exome sequencing. Whole-genome data and targeted sequencing data are also supported when the "coverage" option is specified. If the data you are loading is from a different genome build than your CESAnalysis, you can use the "chain_file" option to supply a UCSC-style chain file, and then your MAF coordinates will be automatically converted with liftOver.

load_maf(
  cesa = NULL,
  maf = NULL,
  annotate = TRUE,
  sample_col = "Tumor_Sample_Barcode",
  chr_col = "Chromosome",
  start_col = "Start_Position",
  ref_col = "Reference_Allele",
  tumor_allele_col = "guess",
  coverage = "exome",
  covered_regions = NULL,
  covered_regions_name = NULL,
  covered_regions_padding = 0,
  progression_col = NULL,
  chain_file = NULL,
  enforce_generic_exome_coverage = FALSE
)

Arguments

cesa
the CESAnalysis object to load the data into

maf
Path of tab-delimited text file in MAF format, or an MAF in data.table or data.frame format

annotate
Annotate mutations with gene and other reference information (required for effect size analysis)

sample_col
column name with sample ID data (Tumor_Sample_Barcode or Unique_Patient_Identifier)

chr_col
column name with chromosome data (Chromosome)

start_col
column name with start position (Start_Position)

ref_col
column name with reference allele data (Reference_Allele)

tumor_allele_col
column name with alternate allele data; by default, values from Tumor_Seq_Allele2 and Tumor_Seq_Allele1 columns are used

coverage
exome, genome, or targeted (default exome)

covered_regions
optional for exome, required for targeted: a GRanges object or a BED file of covered intervals matching the CESAnalysis genome

covered_regions_name
a name describing the covered regions (e.g., "my_custom_targeted_regions"); required when covered_regions are supplied

covered_regions_padding
How many bases (default 0) to expand start and end of each covered_regions interval, to include variants called just outside of targeted regions. Consider setting from 0-100bp, or up to the sequencing read length. If the input data has been trimmed to the targeted regions, leave set to 0.

progression_col
column in MAF with tumor progression state (see docs)

chain_file
a chain file (text format, name ends in .chain) to convert MAF records to the genome build used in the CESAnalysis

enforce_generic_exome_coverage
when loading generic exome data, exclude records that aren't covered in the (somewhat arbitrary) generic exome intervals included with CES genome reference data (default FALSE)

Value

CESAnalysis object with the specified MAF data loaded (in addition to any previously loaded data)